Minor updates to save to json file feature
